assets folder refactor + sidebar styles fix
