+++ src/components/SidebarNavigation/SidebarNav.module.css
@@ -2,7 +2,7 @@
   position: absolute;
   width: 100%;
   height: 100%;
-  z-index: 30;
+  z-index: 9999;
   background: #111518;
   transition: all 0.5s ease;
 }
