+++ src/pages/MainPage/SliderSide/sliderCustomStyles.scss
@@ -18,7 +18,7 @@
     right: 0;
 
     &:after {
-      content: url('../../../assets/images/btnNext.svg');
+      content: url('../../../assets/icons/btnNext.svg');
     }
   }
 
@@ -27,7 +27,7 @@
     left: 0;
 
     &:after {
-      content: url('../../../assets/images/btnPrev.svg');
+      content: url('../../../assets/icons/btnPrev.svg');
     }
   }
